+++ src/components/Account.tsx
@@ -1,8 +1,8 @@
 import { Tab } from "@headlessui/react";
 import { Todo } from "./todo/Todo";
 import { IoBookmarkOutline, IoCheckboxOutline, IoHomeOutline } from "react-icons/io5";
-import { PostList } from "src/components/post/PostList";
-import { Favorite } from "src/components/Favorite";
+import { PostList } from "src/components/home/post/PostList";
+import { Favorite } from "src/components/home/favorite/Favorite";
 import { LoginedLayout } from "src/components/layout/LoginedLayout";
 
 export const Account: React.VFC = () => {
