+++ src/components/layout/GuestLayout.tsx
@@ -1,5 +1,5 @@
-import { Footer } from "../Footer";
-import { Header } from "../Header";
+import { Footer } from "./Footer";
+import { Header } from "./Header";
 
 type Props = {
   children: React.ReactNode;
