+++ src/components/search/SearchDrop.tsx
@@ -15,7 +15,7 @@ export const SearchDrop = () => {
           <Menu.Item>
             {({ active }) => {
               return (
-                <Link href="/postSearch">
+                <Link href="search/postSearch">
                   <a className={`${active && "bg-gray-500 text-white rounded-lg w-20 m-auto"}`}>キロク</a>
                 </Link>
               );
@@ -24,7 +24,7 @@ export const SearchDrop = () => {
           <Menu.Item>
             {({ active }) => {
               return (
-                <Link href="/nameSearch">
+                <Link href="search/nameSearch">
                   <a className={`${active && "bg-gray-500 text-white rounded-lg w-20 m-auto"}`}>ナマエ</a>
                 </Link>
               );
@@ -33,7 +33,7 @@ export const SearchDrop = () => {
           <Menu.Item>
             {({ active }) => {
               return (
-                <Link href="/whSearch">
+                <Link href="search/whSearch">
                   <a className={`${active && "bg-gray-500 text-white rounded-lg w-20 m-auto"}`}>結婚式場</a>
                 </Link>
               );
