+++ src/pages/search/nameSearch.tsx
@@ -1,19 +1,16 @@
 import { useCallback, useState } from "react";
 import { LoginedLayout } from "src/components/layout/LoginedLayout";
 import { SearchForm } from "src/components/search/SearchForm";
-import { SearchUserDisplay } from "src/components/layout/SearchUserDisplay";
+import { SearchUserDisplay } from "src/components/display/SearchUserDisplay";
 import { supabase } from "src/libs/supabase";
 import type { Post } from "./postSearch";
 
-const WhSearch = () => {
+const NameSearch = () => {
   const [posts, setPosts] = useState<Post[]>([]);
 
-  const searchWh = useCallback(async (word: string) => {
+  const searchName = useCallback(async (word: string) => {
     try {
-      const res = await supabase
-        .from("profiles")
-        .select("name,avatar,user_id,wedding_hall")
-        .like("wedding_hall", `%${word}%`);
+      const res = await supabase.from("profiles").select("name,avatar,user_id,wedding_hall").like("name", `%${word}%`);
       if (res.error) throw res.error;
       setPosts(res.data);
     } catch (error) {
@@ -22,10 +19,10 @@ const WhSearch = () => {
   }, []);
   return (
     <LoginedLayout>
-      <SearchForm how={"結婚式場"} handleSearch={searchWh} />
+      <SearchForm how={"ナマエ"} handleSearch={searchName} />
       <SearchUserDisplay posts={posts} />
     </LoginedLayout>
   );
 };
 
-export default WhSearch;
+export default NameSearch;
