+++ src/pages/search/whSearch.tsx
@@ -1,16 +1,19 @@
 import { useCallback, useState } from "react";
 import { LoginedLayout } from "src/components/layout/LoginedLayout";
 import { SearchForm } from "src/components/search/SearchForm";
-import { SearchUserDisplay } from "src/components/layout/SearchUserDisplay";
+import { SearchUserDisplay } from "src/components/display/SearchUserDisplay";
 import { supabase } from "src/libs/supabase";
 import type { Post } from "./postSearch";
 
-const NameSearch = () => {
+const WhSearch = () => {
   const [posts, setPosts] = useState<Post[]>([]);
 
-  const searchName = useCallback(async (word: string) => {
+  const searchWh = useCallback(async (word: string) => {
     try {
-      const res = await supabase.from("profiles").select("name,avatar,user_id,wedding_hall").like("name", `%${word}%`);
+      const res = await supabase
+        .from("profiles")
+        .select("name,avatar,user_id,wedding_hall")
+        .like("wedding_hall", `%${word}%`);
       if (res.error) throw res.error;
       setPosts(res.data);
     } catch (error) {
@@ -19,10 +22,10 @@ const NameSearch = () => {
   }, []);
   return (
     <LoginedLayout>
-      <SearchForm how={"ナマエ"} handleSearch={searchName} />
+      <SearchForm how={"結婚式場"} handleSearch={searchWh} />
       <SearchUserDisplay posts={posts} />
     </LoginedLayout>
   );
 };
 
-export default NameSearch;
+export default WhSearch;
